+++ hutool-core/src/main/java/com/xiaoleilu/hutool/util/CollectionUtil.java
@@ -1429,7 +1429,7 @@ public class CollectionUtil {
 		Collections.sort(list, c);
 		return list;
 	}
-
+	
 	// ------------------------------------------------------------------------------------------------- forEach
 
 	/**